Сделал регистрацию пользователей в серверном гуи. И авторизацию через логин-пароль. Если пользователь не зарегистрирован на сервере, пароль не верен, или пользователь уже авторизовался , то выводятся соответствующие сообщения и можно ввести другие данные для входа.
Сделал декоратор которым обернул функцию старта основного приложения, что бы оно запускалось только если пользователь прошёл авторизацию.

+++ server_storage.py
@@ -10,9 +10,11 @@ from datetime import datetime
 
 class ServerStorage:
     class AllUser:
-        def __init__(self, name, last_login_date):
+        def __init__(self, name, passwrd, last_login_date):
             self.name = name
+            self.passwrd = passwrd
             self.last_login_date = last_login_date
+
             self.id = None
 
         def __repr__(self):
@@ -64,6 +66,7 @@ class ServerStorage:
         clients_table = Table('all_users', self.meta_data,
                               Column('id', Integer, primary_key=True),
                               Column('name', String),
+                              Column('passwrd', String),
                               Column('last_login_date', DateTime))
 
         active_users = Table('active_users', self.meta_data,
@@ -92,16 +95,16 @@ class ServerStorage:
         mapper(self.Contact, contacts)
         self.clear_active_users()
 
-    def create_user(self, name):
-        """Добавляем нового пользователя, если он есть меняем дату последнего входа """
+    def create_user(self, name,passwrd):
+        """Добавляем нового пользователя, если он есть возвращаем False """
         user = self.session.query(self.AllUser).filter_by(name=name).first()
         if user:
-            user.last_login_date = datetime.now()
+            return False
         else:
-            user = self.AllUser(name, datetime.now())
-        self.session.add(user)
-        self.session.commit()
-        return user
+            user = self.AllUser(name,passwrd, datetime.now())
+            self.session.add(user)
+            self.session.commit()
+            return True
 
     def delete_user(self, name):
         """Удаляем пользователя"""
@@ -120,22 +123,28 @@ class ServerStorage:
         return user
 
     def user_login(self, name, ip, port):
-        """ Создает нового пользователя, или меняет дату входа у старого,
+        """
         добавлет в активные и заносит в журнал истории логинов"""
-        new_user = self.create_user(name)
-        self.add_new_active_user(new_user.id, ip, port)
-        history = self.LoginHistory(new_user.id, ip, port)
+        user = self.session.query(self.AllUser).filter_by(name=name).first()
+        self.add_new_active_user(user.id, ip, port)
+        history = self.LoginHistory(user.id, ip, port)
         self.session.add(history)
         self.session.commit()
 
+    def get_user_pass(self,name):
+        user = self.session.query(self.AllUser).filter_by(name=name).first()
+        if user:
+            return user.passwrd
+        else:
+            return False
+
     def delete_active_user(self, user_name):
         ''' Удаляет пользователя из активных '''
-        # user = self.session.query(self.AllUser).filter_by(name=user_name).first()
-        # user = self.session.query(self.ActiveUser).filter_by(user_id=user.id).first()
-        user = \
-            self.session.query(self.AllUser, self.ActiveUser).filter_by(name=user_name).join(self.ActiveUser).first()[1]
-        self.session.delete(user)
-        self.session.commit()
+        if user_name:
+            user = \
+                self.session.query(self.AllUser, self.ActiveUser).filter_by(name=user_name).join(self.ActiveUser).first()[1]
+            self.session.delete(user)
+            self.session.commit()
 
     def clear_active_users(self):
         """ Очищает таблицу активных юзеров"""
@@ -204,32 +213,32 @@ class ServerStorage:
 
 if __name__ == '__main__':
     server = ServerStorage()
-    new_user = server.create_user('Vovas')
+    new_user = server.create_user('Vovas','Paasword')
     server.add_new_active_user(new_user.id, '127.0.0.1', '7777')
-    new_user_2 = server.create_user('Dima')
+    new_user_2 = server.create_user('Dima','Passwrd2')
     server.add_new_active_user(new_user_2.id, '127.0.0.1', '7777')
 
-    server.user_login('Vanya', '127.0.0.1', '8888')
+    # server.user_login('Vanya', '127.0.0.1', '8888')
     full_data = server.session.query(server.AllUser.name, server.ActiveUser.ip, server.AllUser.last_login_date).join(
         server.AllUser).all()
     # print(full_data)
-    server.user_login('Petya', '127.0.0.1', '8888')
+    # server.user_login('Petya', '127.0.0.1', '8888')
     print('-------------')
     print(server.get_active_users(name='Vovas'))
     # server.delete_active_user('Vovas')
-
-    query = server.session.query(server.AllUser, server.ActiveUser).outerjoin(server.ActiveUser)
-    query = server.session.query(server.AllUser, server.ActiveUser).outerjoin(server.ActiveUser)
+    #
+    # query = server.session.query(server.AllUser, server.ActiveUser).outerjoin(server.ActiveUser)
+    # query = server.session.query(server.AllUser, server.ActiveUser).outerjoin(server.ActiveUser)
     # print(query)
-    records = query
+    # records = query
     # for user, active_user in records:
     #     # print(user.name,active_user.ip,active_user.port,user.last_login_date)
     #     print(user.name,
     #           f'Подключился {datetime.strftime(user.last_login_date, "%d.%m.%Y %H:%M")}' if active_user else 'Не активен')
     #     # print(user,active_user)
 
-    resp = server.add_new_contact('User-2', 'User-5')
-    print(resp)
-    resp = server.delete_new_contact('User-2', 'User-5')
-    print(resp)
-    contacts = server.get_contacts('User-17')
+    # resp = server.add_new_contact('User-2', 'User-5')
+    # print(resp)
+    # resp = server.delete_new_contact('User-2', 'User-5')
+    # print(resp)
+    # contacts = server.get_contacts('User-17')
